Move PushRequestInterceptor test code to MainListViewController in FRAuthenticatorExample
